Move Base, _custom_json_serializer from catalog to separate sqlalchemy file
