+++ features/steps/test_inspect_bin.py
@@ -8,7 +8,7 @@ from pytest_bdd import scenario, given, when, then, parsers
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 
-from dor.adapters.catalog import Base, _custom_json_serializer
+from dor.adapters.sqlalchemy import Base, _custom_json_serializer
 from dor.config import config
 from dor.domain.models import Bin
 from dor.providers.models import (
